refactor: front and backend changes
